+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -5,7 +5,7 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
-import woowacourse.shopping.di.getViewModel
+import woowacourse.shopping.ui.common.getViewModel
 
 class CartActivity : AppCompatActivity() {
